feat: Allow a guarantee to keep at least N images
This feature is meant to allow a user to guarantee that they'll always
have at least N number of images. Since this action currently only
operates on time, if you have a service that doesn't update often, you
run the risk of deleting all of your images. Using `keep-at-least`
should give you the ability to keep N number of images. For example,
keep just the latest image.

To keep compatibility with the existing action, the default behavior
that happens if you don't specify `keep-at-least` is to keep 0 images.