feat: Add `use-github-token` as action input
This commit adds the ability to use the built-in GITHUB_TOKEN secret as
passed into the GitHub Actions runner instead of providing a personal
access token (PAT) as a secret into the action for image deletion for a
package that lives in the same repository as the GitHub repository.

In order to do so, `image-name` needs to match the package name for the
repository from where this action is being invoked.