tests: Add test case for bad `token_type` value
This commit adds a test case to validate the `token_type` passed into
Inputs against valid GithubTokenType enum values.

Note to self: In order for pydantic validation to work, the `input_names`
validator needs to check for the existence of `token_type` in
ValidationInput. This will not exist if `token_type` is invalid, and it
will be caught by the default validator. This is similar to what was
done in the `validate_org_name` field validator.